shut off some warnings, more tweaks to autoconf madness
git-svn-id: http://biofs04/svn/bioinf/anfo/trunk@2017 93d9a472-daf7-0310-8f21-d7befa554562

+++ src/martin-table.cc
@@ -192,7 +192,7 @@ template< typename C > istream& read_martin_table_indel( istream& s, C& d, const
 
 		if( r.sequence.size() > 255 )
 			error( 1, 0, "cannot represent long insert at %s:%d (%d)",
-					hsa_chr.c_str(), r.hsa.pos, r.sequence.size() ) ;
+					hsa_chr.c_str(), r.hsa.pos, (int)r.sequence.size() ) ;
 
 		if( type == "deletion" )
 		{
